refactor choice_iter into extract
